+++ src/events/star_spawn.rs
@@ -6,22 +6,33 @@ use rand::Rng;
 #[derive(Debug, Event)]
 pub struct StarSpawnEvent {
     star: Star,
+    id: Option<u64>,
 }
 
 impl StarSpawnEvent {
+    pub fn new(star: Star, id: Option<u64>) -> Self {
+        Self { star, id }
+    }
+
     pub fn get_star(&self) -> &Star {
         &self.star
+    }
+
+    pub fn get_id(&self) -> Option<u64> {
+        self.id
     }
 
     pub fn random(rng: &mut impl Rng, position: Vec3) -> Self {
         Self {
             star: Star::new_random(rng, position),
+            id: None,
         }
     }
 
     pub fn with_mass(position: Vec3, mass: f64) -> Self {
         Self {
             star: Star::new(mass, position),
+            id: None,
         }
     }
 
@@ -33,6 +44,7 @@ impl StarSpawnEvent {
         let mass = stellar_class.generate_random_mass(rng);
         Self {
             star: Star::new(mass, position),
+            id: None,
         }
     }
 }