+++ src/plugins/system/load_on_startup.rs
@@ -1,6 +1,8 @@
-use crate::utils::persistence::{load_state, save_state_exists};
+use crate::events::star_cluster_generate::StarClusterGenerateEvent;
+use crate::persistence::{load_state, save_state_exists};
+use crate::types::cluster_generation_method::ClusterGenerationMethod;
 use bevy::log::info;
-use bevy::prelude::World;
+use bevy::prelude::{Events, World};
 
 pub fn load_on_startup(world: &mut World) {
     if save_state_exists() {