+++ src/utils.rs
@@ -1,4 +1,3 @@
 pub mod base64;
 pub mod misc;
-pub mod persistence;
 pub mod random;